+++ jlrs/src/data/managed/symbol.rs
@@ -1,11 +1,10 @@
 //! Managed type for `Symbol`. Symbols represent identifiers like module and function names.
 
-#[julia_version(windows_lts = false)]
-use std::mem::MaybeUninit;
 use std::{
     ffi::CStr,
     hash::{Hash, Hasher},
     marker::PhantomData,
+    mem::MaybeUninit,
     ptr::NonNull,
 };
 
@@ -13,12 +12,10 @@ use jl_sys::{
     jl_gensym, jl_sym_t, jl_symbol_n, jl_symbol_name_ as jl_symbol_name, jl_symbol_type,
     jl_tagged_gensym,
 };
-use jlrs_macros::julia_version;
 
 use super::Ref;
-#[julia_version(windows_lts = false)]
-use crate::catch::catch_exceptions;
 use crate::{
+    catch::catch_exceptions,
     data::managed::private::ManagedPriv,
     error::{JlrsError, JlrsResult},
     impl_julia_typecheck,
@@ -52,7 +49,6 @@ impl<'scope> Symbol<'scope> {
         }
     }
 
-    #[julia_version(windows_lts = false)]
     /// Convert the given byte slice to a `Symbol`.
     pub fn new_bytes<N, T>(target: T, symbol: N) -> T::Exception<'static, Self>
     where